Update meetup_11_August_2018.sol
Added some comments to try to clarify the course of our exploration. This is not a complete contract in any way and doesn't attempt to be one, were just exploring the technology.
I'm sure I made a bunch of autocorrect spelling mistakes as it's late at night. Feel free to correct them and I'm include that into the repo